requirements.txt
flask==3.0.0
python-gitlab==4.4.0
requests==2.31.0
gunicorn==21.2.0
